move execution of flutter pub outdated into flutterCli.ts;
